r extract interface
