+++ src/main/proxy/ProductService.java
@@ -1,14 +1,5 @@
 package main.proxy;
 
-public class ProductService {
-    private final ProductRepository productRepository;
-
-    public ProductService(final ProductRepository productRepository) {
-        this.productRepository = productRepository;
-    }
-
-    public Product getProductById(final long productId) {
-        return productRepository.findById(productId)
-                .orElseThrow(() -> new IllegalArgumentException("상품을 찾을 수 없습니다."));
-    }
+public interface ProductService {
+    Product getProductById(long productId);
 }
